// types/appserverComponents.ts
export interface ErrMsg {
	ret?: number
	_errno?: number
	_errstr?: string
	_param?: any
}
// ...
export interface NewForgetCodeCheckReq extends ErrMsg {
	acode?: string
	phone?: string
	email?: string // email地址
	code: string // 验证码
}